+++ cli/src/commands/index.ts
@@ -1,4 +1,5 @@
 export * from './generate.command';
 export * from './hash.command';
+export * from './metadata.command';
 export * from './sign.command';
 export * from './verify.command';